+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Fixed
 
+- If a suggestion is selected, the suggestion window will disappear until the next editor change.
 - Separator warning will appear until clicked or the problematic characters are removed, and only instance of this warning will appear at a time.
 
 ## [1.2.0] - 2025-2-25